% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateAccuracy}
\alias{calculateAccuracy}
\title{Calculates classification accuracy by comparing predicted labels with true labels.}
\usage{
calculateAccuracy(y, y_pred)
}
\arguments{
\item{y}{True labels.}

\item{y_pred}{Predicted values.}
}
\value{
Classification accuracy as a percentage.
}
\description{
Calculates classification accuracy by comparing predicted labels with true labels.
}
