% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calL_h_G_prime}
\alias{calL_h_G_prime}
\title{Calculates the first derivative of the Gaussian loss function L_h^G(v).}
\usage{
calL_h_G_prime(v, h)
}
\arguments{
\item{v}{Evaluation point.}

\item{h}{Bandwidth parameter.}
}
\value{
The first derivative of the Gaussian loss function at `v`.
}
\description{
Calculates the first derivative of the Gaussian loss function L_h^G(v).
}
