% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comb}
\alias{comb}
\title{Combine Elements from Multiple Lists}
\usage{
comb(x, ...)
}
\arguments{
\item{x}{First list to combine}

\item{...}{Additional lists to combine with the first list}
}
\value{
A list where each element contains the corresponding elements from all input lists
}
\description{
Utility function for use with foreach to combine elements from multiple lists by position
}
