% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createdir}
\alias{createdir}
\title{Create Directory if Not Exists}
\usage{
createdir(path, recursive = T)
}
\arguments{
\item{path}{Directory path to create.}

\item{recursive}{Create parent directories if needed (default=TRUE).}
}
\value{
Logical: TRUE if directory was created, FALSE if already existed.
}
\description{
Utility function to safely create directories. Equivalent to mkdir -p in Unix.
}
