% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixRNGStream.R
\name{fixRNGStream}
\alias{fixRNGStream}
\title{Fix Random Number Generator Stream}
\usage{
fixRNGStream(batch = 1, seed = 42)
}
\arguments{
\item{batch}{an integer scalar specifying the batch number for the RNG stream.}

\item{seed}{an integer scalar specifying the seed for the RNG.}
}
\value{
The updated RNG stream.
}
\description{
This function sets up the random number generator (RNG) stream for reproducibility.
}
