% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSampleIndices}
\alias{getSampleIndices}
\title{Finds indices of samples that belong to a specific partition.}
\usage{
getSampleIndices(divisions, division)
}
\arguments{
\item{divisions}{Vector indicating which partition each sample belongs to.}

\item{division}{Partition number to extract.}
}
\value{
Armadillo unsigned integer vector containing indices of samples in the specified partition.
}
\description{
Finds indices of samples that belong to a specific partition.
}
