% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_hinge_loss}
\alias{grad_hinge_loss}
\title{Computes the gradient of the hinge loss for a specified kernel type.}
\usage{
grad_hinge_loss(Y, X, beta, h, type)
}
\arguments{
\item{Y}{Response vector.}

\item{X}{Design matrix.}

\item{beta}{Coefficient vector.}

\item{h}{Kernel bandwidth.}

\item{type}{Type of kernel (1–5).}
}
\value{
Gradient of the hinge loss as an Armadillo vector.
}
\description{
Supported types:
- 1: Uniform
- 2: Laplacian
- 3: Logistic
- 4: Gaussian
- 5: Epanechnikov
}
