% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{L2_HingeLoss}
\alias{L2_HingeLoss}
\title{Computes the L2-regularized hinge loss for a specific kernel type.}
\usage{
L2_HingeLoss(v, h, type)
}
\arguments{
\item{v}{Evaluation point.}

\item{h}{Kernel bandwidth.}

\item{type}{Type of kernel (1–5).}
}
\value{
Value of the L2-regularized hinge loss at `v`.
}
\description{
Supported types:
- 1: Uniform
- 2: Laplacian
- 3: Logistic
- 4: Gaussian
- 5: Epanechnikov
}
