% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Phi}
\alias{Phi}
\title{Computes the cumulative distribution function (CDF) of the standard normal distribution.}
\usage{
Phi(x)
}
\arguments{
\item{x}{Input value.}
}
\value{
The value of the CDF at `x`.
}
\description{
This function uses the complementary error function (`std::erfc`) to compute Φ(x).
}
