% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft_thresholding_cpp}
\alias{soft_thresholding_cpp}
\title{Applies soft thresholding operator for L1 regularization.}
\usage{
soft_thresholding_cpp(x, t)
}
\arguments{
\item{x}{Input vector of coefficients.}

\item{t}{Threshold parameter (controls sparsity level).}
}
\value{
Thresholded vector where values within [-t, t] are set to zero.
}
\description{
This implements the proximal operator for L1 norm: S_t(x) = sign(x)*max(|x|-t, 0).
}
